rgb project
This is my projects with my boot camp. it strange how i look at the code and it all clicks as to what Ive done. a giant spider web where this links to that links to this. the layers of abstraction that come with the functions and bindings working together. im looking forward to the next lesson. will update css to improve aperance